Consolidate HashTable with SortedTable
