+++ examples/lib/invoke-function.ts
@@ -14,8 +14,9 @@ const executorService = new Lambda.ExecutorService({
   timeout: cdk.Duration.seconds(10)
 });
 
-const table = new DynamoDB.HashTable(stack, 'my-table', {
+const table = new DynamoDB.Table(stack, 'my-table', {
   partitionKey: 'id',
+  sortKey: undefined,
   shape: {
     id: string(),
     count: integer({
