+++ examples/lib/stream-processing.ts
@@ -29,8 +29,9 @@ const topic = new SNS.Topic(stack, 'Topic', {
 /**
  * Create a DynamoDB Table to store some data.
  */
-const enrichments = new DynamoDB.HashTable(stack, 'Enrichments', {
+const enrichments = new DynamoDB.Table(stack, 'Enrichments', {
   partitionKey: 'key',
+  sortKey: undefined,
   shape: {
     // define the shape of data in the dynamodb table
     key: string(),
@@ -38,6 +39,7 @@ const enrichments = new DynamoDB.HashTable(stack, 'Enrichments', {
   },
   billingMode: BillingMode.PAY_PER_REQUEST
 });
+
 
 /**
  * Schedule a Lambda Function to send a (dummy) message to the SNS topic:
@@ -113,6 +115,7 @@ const stream = queue.stream() // stream gives us a nice chainable API for resour
   .map({
     depends: enrichments.readAccess(),
     handle: async(message, e) => {
+      e
       // here we transform messages received from SQS by looking up some data in DynamoDB
       const enrichment = await e.get({
         key: message.key
