+++ packages/punchcard/lib/storage/dynamodb/index.ts
@@ -1,4 +1,3 @@
-export * from './client';
 export * from './expression';
 export * from './key';
 export * from './query';
